requirements.txt
Flask
Flask-Cors
Flask-Login
Flask-PyMongo
pymongo
python-dotenv==1.0.0